Pasado contenido a script contenido
